feat: 상품의 position과 CartRepository::deleteCartProduct의 id가 동일한 값임을 보장할 수 없다는 문제를 해결한다

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -25,9 +25,9 @@ class CartViewModel(
         }
     }
 
-    fun deleteCartProduct(id: Int) {
+    fun deleteCartProduct(id: Long) {
         viewModelScope.launch {
-            cartRepository.deleteCartProduct(id.toLong())
+            cartRepository.deleteCartProduct(id)
             _onCartProductDeleted.value = true
         }
     }
